+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/hellohttp/ux/TestTag.kt
@@ -56,6 +56,7 @@ enum class TestTagPart {
     EnvironmentDisableSystemCaCertificates,
     ClientCertificate,
     PrivateKey,
+    Bundle,
     CreateButton,
     ListItemLabel,
     Inherited,